Includino documentação e exemplos para requests via swagger ui, incluindo validacoes no usecase para explorar marte e refactory nos pacotes e classes de request e response

+++ src/main/java/gov/nasa/mars/probe/control/entities/Coordinate.java
@@ -4,20 +4,20 @@ import gov.nasa.mars.probe.control.exceptions.InvalidProbeCoordinateException;
 
 public class Coordinate {
 
-	private Integer value;
+	private Long value;
 
 	public Coordinate() {}
 
-	public Coordinate(final Integer value) {
+	public Coordinate(final Long value) {
 		super();
 		this.value = value;
 	}
 
-	public Integer getValue() {
+	public Long getValue() {
 		return value;
 	}
 
-	public void setValue(final Integer value) throws InvalidProbeCoordinateException {
+	public void setValue(final Long value) throws InvalidProbeCoordinateException {
 		
 		if(value < 0)
 			throw new InvalidProbeCoordinateException();
